chore: renamed module dmx to artnet
