+++ src/modules/artnet/consumer/artnet_consumer.h
@@ -19,16 +19,11 @@
 #include "../util/artnet_send.h"
 
 namespace caspar {
-    namespace dmx {
-
-        spl::shared_ptr<core::frame_consumer>
-        create_consumer(const std::vector<std::wstring>&                         params,
-                       const core::video_format_repository&                     format_repository,
-                       const std::vector<spl::shared_ptr<core::video_channel>>& channels);
+    namespace artnet {
 
         spl::shared_ptr<core::frame_consumer>
         create_preconfigured_consumer(const boost::property_tree::wptree&               ptree,
                                                                             const core::video_format_repository&              format_repository,
                                                                             std::vector<spl::shared_ptr<core::video_channel>> channels);
     }
-} // namespace caspar::dmx
+} // namespace caspar::artnet
