+++ src/modules/artnet/util/artnet_send.h
@@ -14,9 +14,9 @@
 #include <vector>
 
 namespace caspar {
-    namespace dmx {
+    namespace artnet {
 
         void send_dmx_data(unsigned short port, const std::wstring& host, int universe, const std::uint8_t* data, size_t length);
         void send_dmx_data(unsigned short port, const std::wstring& host, int universe, std::vector<uint8_t> data);
     }
-} // namespace caspar::dmx
+} // namespace caspar::artnet
