+++ src/shell/casparcg.config
@@ -138,7 +138,7 @@
                 <path>[file|url]</path>
                 <args>[most ffmpeg arguments related to filtering and output codecs]</args>
             </ffmpeg>
-            <dmx>
+            <artnet>
                 <universe>0</universe>
 
                 <host>127.0.0.1</host>
@@ -161,7 +161,7 @@
                         <rotation>0</rotation>
                     </fixture>
                 </fixtures>
-            </dmx>
+            </artnet>
         </consumers>
         <producers>
             <producer id="0">AMB LOOP</producer>
